+++ lib/ecto_network/inet.ex
@@ -8,15 +8,17 @@ defmodule EctoNetwork.INET do
   def type, do: :inet
 
   @doc "Handle casting to Postgrex.INET"
-  def cast(address) do
-    address
-    |> case do
-      %Postgrex.INET{}=address -> {:ok, address}
-      address when is_tuple(address) -> cast(%Postgrex.INET{address: address})
-      address when is_binary(address) -> parse_address(address) |> cast()
-      _ -> :error
+  def cast(%Postgrex.INET{}=address), do: {:ok, address}
+  def cast(address) when is_tuple(address), do: {:ok, %Postgrex.INET{address: address}}
+  def cast(address) when is_binary(address) do
+    case String.split(address, "/") do
+      [address] ->
+        cast_result(parse_address(address), nil)
+      [address, netmask] ->
+        cast_result(parse_address(address), parse_netmask(netmask))
     end
   end
+  def cast(_), do: :error
 
   @doc "Load from the native Ecto representation"
   def load(%Postgrex.INET{}=address), do: {:ok, address}
@@ -27,22 +29,40 @@ defmodule EctoNetwork.INET do
   def dump(_), do: :error
 
   @doc "Convert from native Ecto representation to a binary"
-  def decode(%Postgrex.INET{address: address}) do
+  def decode(%Postgrex.INET{address: address, netmask: nil}) do
     case :inet.ntoa(address) do
       {:error, _einval} -> :error
       formated_address  -> List.to_string(formated_address)
     end
   end
+  def decode(%Postgrex.INET{address: address, netmask: netmask}) do
+    case :inet.ntoa(address) do
+      {:error, _einval} -> :error
+      formatted_address -> List.to_string(formatted_address) <> "/#{netmask}"
+    end
+  end
 
+  @spec parse_address(binary) :: {:ok, :inet.ip_address} | {:error, :einval}
   defp parse_address(address) do
     address
     |> String.to_charlist()
     |> :inet.parse_address()
-    |> case do
-      {:ok, address} -> address
-      {:error, error} -> error
+  end
+
+  @spec parse_netmask(binary) :: integer | :error
+  defp parse_netmask(netmask) do
+    case Integer.parse(netmask) do
+      {value, _rest} -> value
+      :error -> :error
     end
   end
+
+  defp cast_result({:ok, _address}, :error), do: :error
+  defp cast_result({:ok, address}, netmask) do
+    {:ok, %Postgrex.INET{address: address, netmask: netmask}}
+  end
+  defp cast_result(_, _), do: :error
+
 end
 
 defimpl String.Chars, for: Postgrex.INET do